gather all imports on logic
