+++ src/visit/visitActions.ts
@@ -1,6 +1,5 @@
 import { ParsedLogic } from '../types'
 import * as ts from 'typescript'
-import * as path from 'path'
 import { NodeBuilderFlags } from 'typescript'
 import { cloneNode } from '@wessberg/ts-clone-node'
 
@@ -19,22 +18,6 @@ export function visitActions(type: ts.Type, inputProperty: ts.PropertyAssignment
 
         if (ts.isFunctionTypeNode(typeNode)) {
             parameters = signature.getDeclaration().parameters.map((param) => {
-                if (parsedLogic.logicName === 'githubLogic' && name === 'setRepositories') {
-                    // debugger
-                    const type = checker.getTypeAtLocation((param.type as ts.ArrayTypeNode).elementType)
-                    const symbol = type.getSymbol()
-                    const files = symbol.declarations.map(d => d.getSourceFile().fileName).filter(fileName => !fileName.includes('/node_modules/'))
-                    const declaration = symbol.getDeclarations()[0]
-
-                    const isExported = declaration.parent.modifiers[0]?.kind === 92
-
-                    if (isExported && files.length === 1 && files[0] !== parsedLogic.fileName) {
-                        if (!parsedLogic.typeImports[files[0]]) {
-                            parsedLogic.typeImports[files[0]] = new Set()
-                        }
-                        parsedLogic.typeImports[files[0]].add(type.aliasSymbol.escapedName as string)
-                    }
-                }
                 return ts.createParameter(
                     undefined,
                     undefined,
